RF(OPT): move out vcr related functionality under support/vcr_
It shaves off many many milliseconds from "startup" time ;)
